pwsearch/requirements.txt
pymediawiki
rich